+++ webapp/controller/user.js
@@ -3,8 +3,8 @@
  * @tagline         User Controller for jPulse Framework WebApp
  * @description     This is the user controller for the jPulse Framework WebApp
  * @file            webapp/controller/user.js
- * @version         1.6.35
- * @release         2026-03-24
+ * @version         1.6.36
+ * @release         2026-03-25
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
@@ -480,7 +480,8 @@ class UserController {
             }
 
             // Remove sensitive data
-            const { passwordHash, ...userProfile } = user;
+            const { passwordHash, ...restProfile } = user;
+            let userProfile = UserModel.applyExtensionSchemaDefaults(restProfile);
 
             // For non-admin users, remove admin-only fields
             if (!isAdmin) {