+++ webapp/tests/manual/redis/graceful-degradation.sh
@@ -4,8 +4,8 @@
  # @tagline         Test Redis graceful degradation (dev environment only)
  # @description     Tests that the application continues working WITHOUT Redis (single-instance mode)
  # @file            webapp/tests/manual/redis/graceful-degradation.sh
- # @version         1.6.35
- # @release         2026-03-24
+ # @version         1.6.36
+ # @release         2026-03-25
  # @repository      https://github.com/jpulse-net/jpulse-framework
  # @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  # @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/