+++ webapp/utils/metrics-registry.js
@@ -3,8 +3,8 @@
  * @tagline         Centralized registration for component metrics providers
  * @description     Manages registration and discovery of component getMetrics() methods
  * @file            webapp/utils/metrics-registry.js
- * @version         1.6.35
- * @release         2026-03-24
+ * @version         1.6.36
+ * @release         2026-03-25
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/